requirements.txt
django
django-rest-framework 
python-dotenv 
psycopg2-binary
pycryptodome
pyjwt
cryptography
python-jose[cryptography]
colorlog